pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cerf_data_centers"
version = "0.1.0"
description = "Data Center Siting Tool - Identifies optimal locations for data centers based on various constraints and costs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "LICENSE" }
authors = [
    {name = "Kendall Mongird", email = "kendall.mongird@pnnl.gov"},
    {name = "Chris Vernon", email = "chris.vernon@pnnl.gov"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "tqdm>=4.67.1",
    "geopandas>=0.14.2",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "click>=8.1.7",
    "numpy>=2.2.2",
    "rasterio>=1.3.8",
    "shapely>=2.0.3",
    "networkx>=3.3.1",
]

[project.scripts]
cerf-siting = "cerf_data_centers.run_siting:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["cerf_data_centers"]
